+++ src/resources/publicAPI/DMAssetResource.ts
@@ -16,7 +16,7 @@ interface DMAssetResource {
   isUsed: boolean;
   mimetype: string;
   modified: Date;
-  tags: Array<string>;
+  tags: Array<string | any>;
   thumbnails: Array<any>;
   title: string;
   type: string;
@@ -43,7 +43,7 @@ interface DMAssetResource {
  * @prop {Array<object>} thumbnails - Array of all thumbnails
  * @prop {Boolean} isUsed - Whether or not this asses is used in any entry
  * @prop {string} mimetype - Mimetype of the assets file
- * @prop {Array<string>} tags - Array of tags
+ * @prop {Array<string|object>} tags - Array of tags
  */
 class DMAssetResource extends Resource {
   /**
@@ -120,7 +120,9 @@ class DMAssetResource extends Resource {
       tags: {
         enumerable: true,
         get: () => <Array<string>>this.getProperty('tags'),
-        set: (value: Array<string>) => this.setProperty('tags', value),
+        set: (value: Array<string | any>) => {
+          return this.setProperty('tags', value.map((x) => (typeof x === 'string' ? x : x.tag)));
+        },
       },
       thumbnails: {
         enumerable: true,
@@ -165,9 +167,20 @@ class DMAssetResource extends Resource {
    * @returns {Promise<string>} the url string of the requested image
    */
   getFileVariant(size?: number, thumb: boolean = false): Promise<string> {
-    return Promise.resolve().then(() => {
+    return Promise.resolve().then(async () => {
       if (!size && !thumb) {
         return this.file.url;
+      }
+
+      if (!thumb && !this.file.resolution) {
+        return this.file.url;
+      }
+
+      if (!thumb && this.file.resolution) {
+        const biggestDimension = Math.max(this.file.resolution.width, this.file.resolution.height);
+        if (!size || biggestDimension <= size) {
+          return this.file.url;
+        }
       }
 
       let file;
@@ -180,11 +193,11 @@ class DMAssetResource extends Resource {
         return file.url;
       }
 
-      const request = this.newRequest();
+      let request;
       if (thumb) {
-        request.follow('ec:dm-asset/thumbnail');
+        request = await this.follow('ec:dm-asset/thumbnail');
       } else {
-        request.follow('ec:dm-asset/file-variant');
+        request = await this.follow('ec:dm-asset/file-variant');
       }
       const templateParams: any = {};
       if (size) {